rewirte the pidTurn func and add some code to ensure gyro wont fluctuate && we wirte a controller func to let the controller vibrating while gyro is calibrating

+++ LoveEggplant/include/motorControl.h
@@ -1,5 +1,5 @@
 void gyroInit();
-void Move(double lPower, double rPower);
+void Move(float lPower, float rPower);
 void Stop(vex::brakeType type = vex::brake);
 void Lift(int power);
 void Tray(int power, vex::brakeType type = vex::brake);
